+++ Sources/FilterEngine/Schema.swift
@@ -27,20 +27,6 @@ public enum Schema {
     /// it required to write additional migration code in the extension.
     public static let BASE_DIR = ".webext"
 
-    /// UserDefaults key for storing the current schema version.
-    ///
-    /// **IMPORTANT**
-    /// Changing this will lead to invalidating the engine cached in the extension directory and makes
-    /// it required to write additional migration code in the extension.
-    public static let ENGINE_SCHEMA_VERSION_KEY = "com.adguard.safari-converter.schema-version"
-
-    /// UserDefaults key for storing the timestamp when the engine was last built.
-    ///
-    /// **IMPORTANT**
-    /// Changing this will lead to invalidating the engine cached in the extension directory and makes
-    /// it required to write additional migration code in the extension.
-    public static let ENGINE_TIMESTAMP_KEY = "com.adguard.safari-converter.engine-timestamp"
-
     /// Name of the file storing the original, uncompiled filtering rules.
     ///
     /// **IMPORTANT**
@@ -53,6 +39,23 @@ public enum Schema {
 
     /// Name of the file storing the serialized `FilterEngine` index.
     public static let FILTER_ENGINE_INDEX_FILE_NAME = "engine.bin"
+
+    /// Name of the file storing engine meta info (timestamp, schema version)
+    public static let ENGINE_META_FILE_NAME = "meta.bin"
+
+    /// Name of the marker file used to signal that migration (engine rebuild)
+    /// is in progress.
+    ///
+    /// Engine rebuild may happen when the schema version has been changed, but
+    /// the engine files are compiled with the previous version. In this case
+    /// the extension process will try to rebuild the engine using plain text
+    /// rules.
+    ///
+    /// This file is used to detect the situation when the migration was
+    /// launched within the extension process and interrupted due to memory
+    /// constraints. In this case on the next extension launch we will not
+    /// attempt to rebuild the engine again.
+    public static let MIGRATION_MARKER_FILE_NAME = "migration"
 
     /// Name of the lock file used for synchronizing access to shared resources.
     public static let LOCK_FILE_NAME = "lock"